+++ apps/frontend/app/(website)/campaigns/components/InsightsCounter.tsx
