+++ apps/frontend/app/(website)/campaigns/components/SearchBox.tsx
