+++ apps/frontend/app/(website)/campaigns/hooks/useInsightsTabsConfig.tsx
