+++ apps/frontend/app/(website)/campaigns/page.tsx
@@ -1,12 +1,11 @@
 "use client";
 import Tabs, { TabContent } from "@/components/shared/Tabs";
-import type { GithubRepository } from "be-types";
-import { useEffect, useState } from "react";
-import InsightsTable from "./components/InsightsTable";
+import type { GithubRepository } from "@codemod-com/api-types";
+import { useState } from "react";
+import CampaignsTable from "./components/InsightsTable";
 import RepositorySelector from "./components/RepositorySelector";
 import SecondaryHeader from "./components/SecondaryHeader";
 import { useInsightTabsConfig } from "./hooks/useInsightsTabsConfig";
-import { useWorkflowRunArtifacts } from "./hooks/useWorflowRunArtifacts";
 import { useCreateWorkflowRun } from "./hooks/useWorkflowRun";
 import { useWorkflowRun } from "./hooks/useWorkflowRunData";
 
@@ -20,33 +19,34 @@ const InsightsPage = () => {
 
   const workflowRun = useWorkflowRun(workflowRunId ?? "");
 
-  const { data } = useWorkflowRunArtifacts(
-    workflowRunId ?? "",
-    workflowRun?.state === "done",
-  );
+  // const { data } = useWorkflowRunArtifacts(
+  //   workflowRunId ?? "",
+  //   workflowRun?.state === "done",
+  // );
 
-  const insights = data?.data?.[0] ?? null;
+  // const insights = data?.data?.[0] ?? null;
 
-  useEffect(() => {
-    if (repo === null) {
-      return;
-    }
+  // useEffect(() => {
+  //   if (repo === null) {
+  //     return;
+  //   }
 
-    onWorkflowRunStart(repo);
-  }, [repo, onWorkflowRunStart]);
+  //   onWorkflowRunStart(repo);
+  // }, [repo, onWorkflowRunStart]);
 
-  useEffect(() => {
-    if (!repo) {
-      setOpen(true);
-    }
-  }, [repo]);
+  // useEffect(() => {
+  //   if (!repo) {
+  //     setOpen(true);
+  //   }
+  // }, [repo]);
 
   return (
     <>
       <div className="relative flex flex-col items-center justify-center">
         <SecondaryHeader
           onOpenRepoSelector={() => setOpen(true)}
-          insightsCount={insights?.migrations.length ?? 0}
+          // insightsCount={insights?.migrations.length ?? 0}
+          insightsCount={0}
           selectedRepoName={repo?.name ?? "Select Repository"}
         />
         <div className="flex w-full justify-start">
@@ -57,17 +57,18 @@ const InsightsPage = () => {
               listClassName="h-[28px]"
               items={tabsConfig.map(({ kind, label }) => ({ id: kind, label }))}
             >
-              {workflowRun?.state === "queued" ||
+              {/* {workflowRun?.state === "queued" ||
                 (workflowRun?.state === "in_progress" &&
                   "Generating insights...")}
-              {workflowRun?.state === "errored" && "Something went wrong..."}
-              {workflowRun?.state === "done" && insights
-                ? tabsConfig.map(({ kind }) => (
-                    <TabContent forId={kind}>
-                      <InsightsTable insights={insights} type={kind} />
-                    </TabContent>
-                  ))
-                : null}
+              {workflowRun?.state === "errored" && "Something went wrong..."} */}
+              {/* {workflowRun?.state === "done" && insights
+                ?  */}
+              {tabsConfig.map(({ kind }) => (
+                <TabContent forId={kind}>
+                  <CampaignsTable insights={[]} type={kind} />
+                </TabContent>
+              ))}
+              {/* : null} */}
             </Tabs>
           )}
         </div>
